requirements.txt
amqp==5.2.0
billiard==4.2.0
blinker==1.8.2
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
kombu==5.3.7
MarkupSafe==2.1.5
prompt_toolkit==3.0.47
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.3
